fix(agent): change to a secondary chroot_exec so it happens in a different process
Due to distroless we can't do a subprocess.run(chroot /root [cmd])
So instead we call a separate python process to chroot inside of. If
we do the chroot inside controller the chroot last for the rest of the
process which causes issues with file management and other things that
happen after the call.