feat: fix agent for distroless and have scr name in flag/history/log (#65)
feat(agent): change to a secondary chroot_exec so it happens in a different process

Due to distroless we can't do a subprocess.run(chroot /root [cmd])
So instead we call a separate python process to chroot inside of. If
we do the chroot inside controller the chroot last for the rest of the
process which causes issues with file management and other things that
happen after the call.

feat(agent): allow passing of root dirs for flag/history/log

For GKE we need to be able to control where flag and history files go
so that they persist between reboots when using the ContainerOS which
only as a few specific stateful directories that are writeable.

feat(operator): set env vars for package/interrupt that have SCR name

In order to disambiguate between the same package in different SCRs
the operator will add the SCR name to the root directories passed to the
agent.

feat: add test automation for full integration test between operator and agent